requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
claude-agent-sdk==0.1.18
click==8.3.1
croniter==6.0.0
cryptography==46.0.3
discord-py==2.6.4
distro==1.9.0
firecrawl-py==4.12.0
frozenlist==1.8.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mcp==1.25.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mem0ai==1.0.1
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.4.0
openai==2.14.0
platformdirs==4.5.1
portalocker==3.2.0
posthog==7.4.3
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
qdrant-client==1.16.2
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
six==1.17.0
slack-bolt==1.27.0
slack-sdk==3.39.0
sniffio==1.3.1
sqlalchemy==2.0.45
sse-starlette==3.1.2
starlette==0.50.0
textual==7.0.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uc-micro-py==1.0.3
urllib3==2.6.2
uvicorn==0.40.0
websockets==15.0.1
yarl==1.22.0
